🚚 rename: 잘못된 경로 수정
