+++ seoyoung059/Week_27/BOJ_1158/BOJ_1158.md
@@ -1,0 +1,39 @@
+## 풀이과정
+- 리스트에 1부터 n까지의 수를 모두 넣고, 규칙에 따라 idx를 바꿔주면서 remove해주면 된다.
+- 처음에 배열에서 모두 체크했는데 시간이 2000ms가 넘게 나왔다....
+
+## 코드
+```java
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.StringTokenizer;
+
+public class Main {
+    public static void main(String[] args) throws IOException {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+        StringTokenizer st = new StringTokenizer(br.readLine());
+
+        int n = Integer.parseInt(st.nextToken());
+        int k = Integer.parseInt(st.nextToken());
+
+        List<Integer> list=new ArrayList<>(n);
+        for (int i = 0; i < n; i++) {
+            list.add(i+1);
+        }
+
+        StringBuilder sb = new StringBuilder();
+        sb.append('<');
+        for (int i = (k-1)%n; list.size()>1; i = (i+k-1)%list.size()) {
+            sb.append(list.remove(i)).append(", ");
+        }
+        sb.append(list.remove(0));
+        sb.append('>');
+
+
+        System.out.println(sb);
+    }
+}
+```
